+++ packages/dashboard/vite.config.ts
@@ -1,17 +1,19 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 
+const port = process.env.PORT || 40000;
+
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [react()],
   server: {
     proxy: {
       '/api': {
-        target: 'http://127.0.0.1:3000',
+        target: `http://127.0.0.1:${port}`,
         changeOrigin: true
       },
       '/ws': {
-        target: 'ws://127.0.0.1:3000',
+        target: `ws://127.0.0.1:${port}`,
         ws: true
       }
     }